# protocol/openapi/agent-v1.yaml
{
  "openapi": "3.1.0",
  "info": {"title": "unYOLO Agent Operations API", "version": "1.0.0"},
  "paths": {
    "/.well-known/unyolo-agent": {
      "get": {"operationId": "discoverAgent", "security": [{"agentBearer": []}], "responses": {"200": {"description": "Agent Operations V1 discovery", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Descriptor"}}}}, "default": {"$ref": "#/components/responses/ErrorResponse"}}}
    },
    "/api/agent/v1/operations": {
      "get": {
        "operationId": "listAgentOperations", "security": [{"agentBearer": []}],
        "parameters": [
          {"name": "idempotency_key", "in": "query", "schema": {"$ref": "#/components/schemas/IdempotencyKey"}},
          {"name": "state", "in": "query", "schema": {"$ref": "#/components/schemas/State"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 50, "default": 20}},
          {"name": "cursor", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 128}}
        ],
        "responses": {"200": {"description": "Newest requester-owned operations", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OperationPage"}}}}, "default": {"$ref": "#/components/responses/ErrorResponse"}}
      },
      "post": {
        "operationId": "submitAgentOperation", "security": [{"agentBearer": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubmitRequest"}}}},
        "responses": {"200": {"description": "Existing or completed operation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Operation"}}}}, "202": {"description": "Pending operation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Operation"}}}}, "default": {"$ref": "#/components/responses/ErrorResponse"}}
      }
    },
    "/api/agent/v1/operations/{id}": {
      "get": {"operationId": "getAgentOperation", "security": [{"agentBearer": []}], "parameters": [{"$ref": "#/components/parameters/OperationID"}], "responses": {"200": {"description": "Agent operation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Operation"}}}}, "default": {"$ref": "#/components/responses/ErrorResponse"}}}
    },
    "/api/agent/v1/operations/{id}/events": {
      "get": {
        "operationId": "waitForAgentOperation", "security": [{"agentBearer": []}],
        "parameters": [{"$ref": "#/components/parameters/OperationID"}, {"name": "after_revision", "in": "query", "schema": {"type": "integer", "minimum": 0}}, {"name": "wait_seconds", "in": "query", "schema": {"type": "integer", "minimum": 0, "maximum": 30}}],
        "responses": {"200": {"description": "Current operation after a revision change or timeout", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Operation"}}}}, "default": {"$ref": "#/components/responses/ErrorResponse"}}
      }
    },
    "/api/agent/v1/operations/{id}/cancel": {
      "post": {
        "operationId": "cancelAgentOperation", "security": [{"agentBearer": []}],
        "parameters": [{"$ref": "#/components/parameters/OperationID"}],
        "responses": {"200": {"description": "Canceled or already terminal operation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Operation"}}}}, "default": {"$ref": "#/components/responses/ErrorResponse"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"agentBearer": {"type": "http", "scheme": "bearer"}},
    "parameters": {"OperationID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}}},
    "responses": {"ErrorResponse": {"description": "Stable error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}},
    "schemas": {
      "Descriptor": {
        "type": "object", "additionalProperties": false,
        "required": ["api_version", "contract_digest", "build_id", "operations", "credential"],
        "properties": {
          "api_version": {"type": "string", "const": "unyolo.io/agent/v1"},
          "contract_digest": {"type": "string", "pattern": "^sha256:[0-9a-f]{64}$"},
          "build_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "operations": {"type": "array", "maxItems": 4096, "uniqueItems": true, "items": {"type": "string", "minLength": 1, "maxLength": 128}},
          "credential": {"$ref": "#/components/schemas/CredentialDescriptor"}
        }
      },
      "CredentialDescriptor": {
        "type": "object", "additionalProperties": false,
        "required": ["ready", "provider", "credential_kind", "generation", "verification_state"],
        "properties": {
          "ready": {"type": "boolean"},
          "provider": {"type": "string", "maxLength": 64},
          "credential_kind": {"type": "string", "maxLength": 64},
          "generation": {"type": "integer", "minimum": 0},
          "verification_state": {"type": "string", "maxLength": 32}
        }
      },
      "State": {"type": "string", "enum": ["pending", "approved", "executing", "succeeded", "failed", "denied", "expired", "canceled"]},
      "Presentation": {"type": "object", "additionalProperties": false, "required": ["title"], "properties": {"title": {"type": "string", "minLength": 1, "maxLength": 160}, "summary": {"type": "string", "maxLength": 500}}},
      "OperationError": {"type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": {"code": {"type": "string", "minLength": 1, "maxLength": 64}, "message": {"type": "string", "minLength": 1, "maxLength": 500}}},
      "IdempotencyKey": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"},
      "SubmitRequest": {"type": "object", "additionalProperties": false, "required": ["idempotency_key", "operation", "target", "arguments", "reason"], "properties": {"idempotency_key": {"$ref": "#/components/schemas/IdempotencyKey"}, "operation": {"type": "string", "minLength": 1, "maxLength": 128}, "target": {"type": "object", "maxProperties": 32}, "arguments": {"type": "object", "maxProperties": 128}, "reason": {"type": "string", "minLength": 1, "maxLength": 2000}}},
      "Operation": {
        "type": "object", "additionalProperties": false,
        "required": ["api_version", "id", "broker", "client_id", "idempotency_key", "operation", "target", "arguments", "reason", "state", "revision", "created_at", "updated_at", "presentation"],
        "properties": {
          "api_version": {"type": "string", "const": "unyolo.io/agent/v1"}, "id": {"type": "string", "minLength": 1, "maxLength": 128}, "broker": {"type": "string", "minLength": 1, "maxLength": 64}, "client_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "idempotency_key": {"$ref": "#/components/schemas/IdempotencyKey"}, "operation": {"type": "string", "minLength": 1, "maxLength": 128}, "target": {"type": "object", "maxProperties": 32}, "arguments": {"type": "object", "maxProperties": 128}, "reason": {"type": "string", "minLength": 1, "maxLength": 2000},
          "state": {"$ref": "#/components/schemas/State"}, "revision": {"type": "integer", "minimum": 1}, "approval_id": {"type": "string", "minLength": 1, "maxLength": 128}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}, "terminal_at": {"type": "string", "format": "date-time"},
          "presentation": {"$ref": "#/components/schemas/Presentation"}, "plan_digest": {"type": "string", "pattern": "^[0-9a-f]{64}$"}, "result": {"type": "object", "maxProperties": 128}, "error": {"$ref": "#/components/schemas/OperationError"}
        }
      },
      "OperationPage": {
        "type": "object", "additionalProperties": false,
        "required": ["api_version", "operations", "next_cursor"],
        "properties": {
          "api_version": {"type": "string", "const": "unyolo.io/agent/v1"},
          "operations": {"type": "array", "maxItems": 50, "items": {"$ref": "#/components/schemas/OperationSummary"}},
          "next_cursor": {"type": ["string", "null"], "minLength": 1, "maxLength": 128}
        }
      },
      "OperationSummary": {
        "type": "object", "additionalProperties": false,
        "required": ["api_version", "id", "broker", "client_id", "idempotency_key", "operation", "state", "revision", "created_at", "updated_at", "presentation"],
        "properties": {
          "api_version": {"type": "string", "const": "unyolo.io/agent/v1"}, "id": {"type": "string", "minLength": 1, "maxLength": 128}, "broker": {"type": "string", "minLength": 1, "maxLength": 64}, "client_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "idempotency_key": {"$ref": "#/components/schemas/IdempotencyKey"}, "operation": {"type": "string", "minLength": 1, "maxLength": 128}, "state": {"$ref": "#/components/schemas/State"}, "revision": {"type": "integer", "minimum": 1},
          "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}, "terminal_at": {"type": "string", "format": "date-time"}, "presentation": {"$ref": "#/components/schemas/Presentation"}
        }
      },
      "ErrorEnvelope": {"type": "object", "additionalProperties": false, "required": ["error"], "properties": {"error": {"$ref": "#/components/schemas/OperationError"}}}
    }
  }
}